requirements.txt
absl-py==1.0.0
addict==2.4.0
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
black==21.12b0
bleach==4.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
configparser==5.2.0
cycler==0.11.0
Cython==0.29.28
dataclasses==0.6
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
docker-pycreds==0.4.0
docopt==0.6.2
entrypoints==0.3
executing==0.8.2
fonttools==4.28.5
future==0.18.2
fvcore==0.1.5.post20220212
gitdb==4.0.9
GitPython==3.1.26
google-auth==2.3.3
google-auth-oauthlib==0.4.6
googledrivedownloader==0.4
grpcio==1.43.0
h5py==2.9.0
idna==3.3
imageio==2.16.2
importlib-metadata==4.10.1
importlib-resources==5.4.0
iopath==0.1.9
ipykernel==6.6.1
ipython==8.0.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyter-packaging==0.11.1
jupyter-server==1.13.2
jupyterlab==3.2.7
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lpips==0.1.4
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
notebook==6.4.7
numpy==1.22.0
#oauthlib==3.1.1
#open3d==0.13.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pathtools==0.1.2
pcl==0.0.0.post1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
pipreqs==0.4.10
platformdirs==2.4.1
plotly==5.6.0
portalocker==2.4.0
prometheus-client==0.12.0
promise==2.3
prompt-toolkit==3.0.24
protobuf==3.19.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
pyntcloud==0.1.6
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
rdflib==6.1.1
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
sentry-sdk==1.5.2
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
stack-data==0.1.3
subprocess32==3.5.4
tabulate==0.8.9
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2022.4.8
tim==0.5
timm==0.4.12
toml==0.10.2
tomli==1.2.3
tomlkit==0.8.0
torch-geometric==2.0.3
torch-tb-profiler==0.3.1
torchsummary==1.5.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
transforms3d==0.3.1
typing-extensions==4.0.1
urllib3==1.26.8
wandb==0.12.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.2
widgetsnbextension==3.5.2
yacs==0.1.8
yarg==0.1.9
yaspin==2.1.0
zipp==3.7.0
